+++ Cargo.toml
@@ -28,5 +28,7 @@ sui_sdk = { git = "https://github.com/mystenlabs/sui", rev = "42ba6c0", package 
 shared_crypto = { git = "https://github.com/MystenLabs/sui", rev = "42ba6c0", package = "shared-crypto" }
 move-core-types = { git = "https://github.com/MystenLabs/sui.git", rev = "42ba6c0" }
 
+sui-sdk-types = "0.0.6"
+
 [profile.release]
 panic = 'abort'
