+++ crates/key-server/Cargo.toml
@@ -5,10 +5,6 @@ authors.workspace = true
 edition.workspace = true
 license.workspace = true
 
-[lib]
-name = "key_server"
-path = "src/lib.rs"
-
 [[bin]]
 name = "key-server"
 path = "src/server.rs"
@@ -32,6 +28,7 @@ tokio = { version = "1.45.1", features = ["full"] }
 axum = { version = "0.7", features = ["macros"] }
 tower-http = { version = "0.6.6", features = ["cors"] }
 crypto = { path = "../crypto" }
+seal-sdk = { path = "../seal-sdk" }
 tap = "1.0.1"
 prometheus = "0.13.3"
 anyhow = "1.0.79"
