+++ crates/key-server/src/tests/externals.rs
@@ -5,7 +5,7 @@ use crate::errors::InternalError;
 use crate::time::current_epoch_time;
 use crate::valid_ptb::ValidPtb;
 use crate::{
-    signed_message,
+    signed_message, signed_request,
     types::{ElGamalPublicKey, ElgamalVerificationKey},
     Certificate, Server,
 };
@@ -13,7 +13,6 @@ use crypto::elgamal;
 use fastcrypto::ed25519::Ed25519Signature;
 use fastcrypto::traits::{KeyPair, Signer};
 use fastcrypto::{ed25519::Ed25519KeyPair, groups::bls12381::G1Element};
-use key_server::signed_request;
 use rand::thread_rng;
 use shared_crypto::intent::{Intent, IntentMessage, PersonalMessage};
 use sui_types::{
