+++ crates/seal-sdk/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "seal-sdk"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+chrono = "0.4.39"
+serde = { workspace = true }
+fastcrypto = { workspace = true }
+bcs = { workspace = true }
+tracing = { workspace = true }
+crypto = { path = "../crypto" }
+sui_types = { workspace = true}
+
+[dev-dependencies]
+rand = { workspace = true }
+hex = { workspace = true }
