+++ crates/seal-sdk/src/lib.rs
@@ -3,6 +3,16 @@
 
 pub mod types;
 
+// Re-export crypto types and functions
+pub use crypto::elgamal::{decrypt as elgamal_decrypt, genkey, SecretKey as ElGamalSecretKey};
+pub use crypto::ibe::PublicKey as IBEPublicKey;
+pub use crypto::{
+    seal_decrypt, seal_encrypt, EncryptedObject, EncryptionInput, IBEPublicKeys, IBEUserSecretKeys,
+};
+
+// Re-export types module contents
+pub use types::{Certificate, DecryptionKey, FetchKeyRequest, FetchKeyResponse, KeyId};
+
 use crate::types::{ElGamalPublicKey, ElgamalVerificationKey};
 use chrono::{DateTime, Utc};
 use fastcrypto::ed25519::Ed25519PublicKey;
@@ -10,9 +20,8 @@ use serde::{Deserialize, Serialize};
 use sui_types::transaction::ProgrammableTransaction;
 use tracing::debug;
 
-/// The format of the personal message shown to the user.
 pub fn signed_message(
-    package_name: String, // should use the original package id
+    package_name: String,
     vk: &Ed25519PublicKey,
     creation_time: u64,
     ttl_min: u16,
@@ -21,7 +30,7 @@ pub fn signed_message(
         "Accessing keys of package {} for {} mins from {}, session key {}",
         package_name,
         ttl_min,
-        DateTime::<Utc>::from_timestamp((creation_time / 1000) as i64, 0) // convert to seconds
+        DateTime::<Utc>::from_timestamp((creation_time / 1000) as i64, 0)
             .expect("tested that in the future"),
         vk,
     );
@@ -69,7 +78,7 @@ mod tests {
             ObjectID::from_str("0xc457b42d48924087ea3f22d35fd2fe9afdf5bdfe38cc51c0f14f3282f6d5")
                 .unwrap();
         let (_, kp): (_, Ed25519KeyPair) = deterministic_random_account_key();
-        let creation_time = 1622548800; // Fixed timestamp
+        let creation_time = 1622548800;
         let ttl_min = 30;
 
         let expected_output = "Accessing keys of package 0x0000c457b42d48924087ea3f22d35fd2fe9afdf5bdfe38cc51c0f14f3282f6d5 for 30 mins from 1970-01-19 18:42:28 UTC, session key DX2rNYyNrapO+gBJp1sHQ2VVsQo2ghm7aA9wVxNJ13U=";
@@ -86,7 +95,7 @@ mod tests {
     #[test]
     fn test_signed_message_mvr_regression() {
         let (_, kp): (_, Ed25519KeyPair) = deterministic_random_account_key();
-        let creation_time = 1622548800; // Fixed timestamp
+        let creation_time = 1622548800;
         let ttl_min = 30;
 
         let expected_output = "Accessing keys of package @my/package for 30 mins from 1970-01-19 18:42:28 UTC, session key DX2rNYyNrapO+gBJp1sHQ2VVsQo2ghm7aA9wVxNJ13U=";
